refactor: about plugin
